chore(ci): ruff + pytest + eslint + vitest ejecutándose en GitHub Actions
Cierra Prioridad 2 del plan del debt audit. El pipeline ahora bloquea
el deploy si falla cualquier check.

Backend:
- Nuevo back/pyproject.toml con ruff (E, F, I, UP, B, RUF, SIM, T10) y
  pytest configurado (strict markers, filtros de warnings de passlib).
- Auto-fix de 22 issues + 8 manuales (imports fuera de orden, líneas con
  múltiples statements, local var sin uso en correcciones.py:634, int()
  redundante sobre round()). Zero warnings.
- Nuevo back/tests/test_api_smoke.py con 6 tests: /health, /openapi.json,
  protegido-sin-token, login V3 schema, login legacy schema. 10 tests totales
  pasando en 0.5s.

Frontend:
- Nuevo front/eslint.config.js (ESLint 9 flat config) con typescript-eslint,
  react-hooks, react-refresh. Reglas degradadas a 'warn' para no romper el
  build — el codebase actual acumula 50 warnings de react-hooks que se
  limpiarán en un sweep separado. package.json ajusta "lint" a --max-warnings=150.
- Nuevos devDeps: typescript-eslint, eslint-plugin-react-hooks,
  eslint-plugin-react-refresh, globals.
- Fix de test_catalogos_mappers: el test esperaba tipos-fcoop/figuras-cooperadoras
  dentro del grupo "Catalogos" pero viven en USER_NAVIGATION; alineado al
  estado real del código.

CI (.github/workflows/deploy.yml):
- Nuevos pasos antes del deploy: ruff check, pytest, npm run lint, npm run test:unit.
- Deploy se bloquea si cualquier check falla.

+++ back/app/api/routes/legacy/informe_general.py
@@ -257,7 +257,7 @@ def _compute_trampeo(session: Session, params: dict, agregado: list[dict] | None
     # (cada TMIMF 'O' es un huerto × semana con sus trampas).
     huertos_count = len(agregado)
     trampas_instaladas_total = (
-        int(round(trampas_instaladas_acumulado / semanas)) if semanas > 0 and huertos_count > 0 else 0
+        round(trampas_instaladas_acumulado / semanas) if semanas > 0 and huertos_count > 0 else 0
     )
     trampas_x_semanas = trampas_instaladas_acumulado
